Changes some data types and casts so there are no extra warnings. Simplified user.c working

+++ include/sched.h
@@ -18,9 +18,9 @@ enum state_t { ST_RUN, ST_READY, ST_BLOCKED, ST_ZOMBIE };
 struct task_struct {
   int PID;			/* Process ID */
   page_table_entry * dir_pages_baseAddr;
-  unsigned int *kernel_esp;
+  unsigned long *kernel_esp;
   int quantum;
-	enum state_t state;
+  enum state_t state;
   struct list_head list;
 };
